requirements/dev.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
djangorestframework==3.12.4
flake8==4.0.1
idna==3.3
kombu==5.2.2
mccabe==0.6.1
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
pytz==2021.3
redis==4.0.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
